lambda-cost-reporting-system/cdk-infrastructure/infrastructure/requirements.txt
aws-cdk-lib==2.122.0
constructs>=10.3.0